final/requirements.txt
pandas==2.0.3
tabulate==0.9.0
nltk==3.8.1
scikit-learn==1.3.1
s3fs==2023.10.0
matplotlib==3.7.0
tk-tools==0.16.0
apache-airflow==2.7.3
numpy~=1.26.1
SQLAlchemy~=1.4.50
requests~=2.31.0